### 2025-02-27 #7813 @alessandrofelder
>would you be able to review this submission to JOSS?

Yes, happy to.
