### 2025-03-07 #7882 @editorialbot
[REVIEW]: ez-zarr: A Python package for easy access and visualisation of OME-Zarr filesets
**Submitting author:** @mbstadler (<a href="http://orcid.org/0000-0002-2269-4934">Michael B. Stadler</a>)
**Repository:** https://github.com/fmicompbio/ez_zarr
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.8
