### 2025-03-10 fmicompbio/ez_zarr#29 @alessandrofelder
Test across OSs and/or document what OS are supported
(With apologies if I've missed anything)

I haven't seen anywhere in the docs what OSs `ez-zarr` supports - given it's pure Python and aimed an novice/intermediate Python users, I presume the aim is to support Windows and Mac too. This should be documented.

However, tests currently only run on Ubuntu.
Should Mac and Windows be tested in CI too (maybe just on one Python version to reduce CI usage)?

Related to https://github.com/openjournals/joss-reviews/issues/7882
