### 2025-03-10 fmicompbio/ez_zarr#30 @alessandrofelder
Explain HCS and Legacy wrappers in docs
(Again, many apologies if I missed this)

From reading the docs, its unclear to me why the HCS Legacy classes exist and when, as I user, I should use them instead of the `ez_zarr.ome_zarr.Image` class?

Could this be improved in the docs?

related to https://github.com/openjournals/joss-reviews/issues/7882
