### 2025-03-10 #7882 @alessandrofelder
Some suggestions for paper text improvements (cc @mbstadler) - very happy with the paper in general - thanks!
* [ ] should https://github.com/bioio-devs/bioio-ome-zarr be cited, and differences to `ez-zarr` explained?
* [ ] (probably in both docs and paper) should there be explicit mention around which versions of zarr (v2 and/or v3?) and OME-zarr are currently supported by `ez-zarr` and how you envisage moving along with the versions?
* [ ] should it be made explicit (again in docs and paper) whether `ez-zarr` supports OME-zarr images with a significant third spatial dimension, and/or temporal dimension, too, or not? (This is unclear to me - the examples and argument names like `upper_left_yx` all seem to assume essentially 2D (maybe z-projected) microscopy images, but the paper seems to suggest support for any OME-zarr images?)
