### 2025-03-13 31b6dce
add bioio-ome-zarr citation
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
