+++ paper/paper.bib
@@ -124,6 +124,13 @@
   year         = 2025
 }
 
+@MISC{bioio,
+  title        = "{bioio-ome-zarr}",
+  author       = "{The BioIO developers}",
+  howpublished = "\url{https://pypi.org/project/bioio-ome-zarr/}",
+  year         = 2025
+}
+
 @MISC{ngio,
   title        = "{ngio}",
   author       = "Cerrone, Lorenzo",
