### 2025-03-13 362a395
add statements on supported zarr versions
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
