### 2025-03-13 d365a72
clarify that OME-Zarr files with any number of dimensions are supported
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
