+++ paper/paper.md
@@ -43,7 +43,7 @@ bibliography: paper.bib
 
 Imaging techniques are an important and growing source of scientific data, but the many different standards and file formats for representing and storing such data make streamlined analysis and data sharing cumbersome. To address this issue, the imaging community has recently developed an open and flexible file format specification (OME-NGFF) [@Moore2021-dw] and a concrete implementation using the Zarr file format (OME-Zarr) [@Moore2023-fv], which is increasingly adopted as the way to store and share multidimensional bioimaging data.
 
-The aim of `ez-zarr` (pronounced “easy zarr”) is to provide high-level, read-only programmatic access to OME-Zarr files, abstracting away the complex, nested representation of these on disk. Internally, the OME-Zarr files are represented using a custom python class, which standardises simple operations such as metadata extraction, image subsetting and visualisation. In addition, the underlying image data, or subsets thereof, can be extracted as `numpy` arrays [@harris2020-np], which makes `ez-zarr` a valuable building block in more complex image analysis workflows that make use of tools that may not yet support OME-Zarrs. 
+The aim of `ez-zarr` (pronounced “easy zarr”) is to provide high-level, read-only programmatic access to OME-Zarr files, abstracting away the complex, nested representation of these on disk. Internally, the OME-Zarr files are represented using a custom python class, which standardises simple operations such as metadata extraction, image subsetting and visualisation. OME-Zarr files with any number of dimensions are supported, and images with more than two spatial dimensions are automatically projected for plotting. In addition, the underlying image data, or subsets thereof, can be extracted as `numpy` arrays [@harris2020-np], which makes `ez-zarr` a valuable building block in more complex image analysis workflows that make use of tools that may not yet support OME-Zarrs. 
 
 `ez-zarr` also facilitates working with collections of OME-Zarr files, such as images from high-content screens where each microwell in a plate is represented by one OME-Zarr file and the OME-Zarr plate collection provides necessary metadata. Such plate-based high-content screening experiments are becoming increasingly popular in research, and entire analysis workflows, like Fractal [@Fractal], now streamline the generation and processing of OME-Zarr datasets.
 
