### 2025-03-13 9fc1bf2
also test on macOS and windows
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
