### 2025-03-13 557670b
normalize paths for windows compatibility
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
