### 2025-03-13 687c5f3
recommend using ome_zarr over hcs_wrappers
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
