+++ docs/getting_started_hcs_wrappers.qmd
@@ -5,6 +5,10 @@ author: Silvia Barbiero, Michael Stadler
 toc: true
 ---
 
+# Disclaimer
+
+The `hcs_wrappers` module described here is only provided for legacy reasons and not being further developed. For new projects, we recommend to use the `ome_zarr` module instead (see [getting started with individual images](https://fmicompbio.github.io/ez_zarr/getting_started_Image/)).
+
 # Goal
 
 The aim of `ez_zarr` is to provide easy, high-level access to high content screening microscopy data, stored as OME-Zarr filesets according to the [NGFF](https://ngff.openmicroscopy.org/latest/) specifications with additional metadata fields, for example the ones generated by the [Fractal](https://fractal-analytics-platform.github.io/) platform.
