### 2025-03-13 fmicompbio/ez_zarr#29 @mbstadler
Good point! We have added `macOS-latest` and `windows-latest` to the tested platforms (mostly in https://github.com/fmicompbio/ez_zarr/commit/9fc1bf298990219c3ea873096b0024909926887c).
