### 2025-03-13 fmicompbio/ez_zarr#29 @alessandrofelder
Nice!
Maybe also worth mentioning cross-OS support explicitly in the docs?
