### 2025-03-13 fmicompbio/ez_zarr#30 @alessandrofelder
Awesome!
