### 2025-03-14 fmicompbio/ez_zarr#29 @mbstadler
Agreed. We have added it in https://github.com/fmicompbio/ez_zarr/commit/290233ab2e5277cc20ae757d55780e09b051a81e
