### 2025-03-16 3d41c47
Select a single timepoint (and remove time axis) for plotting
