### 2025-03-16 82bafda
remove time axis also from labels
