### 2025-03-21 fmicompbio/ez_zarr#37 @pr4deepr
opening 3D+time microscopy dataset
I have a lattice lightsheet dataset, which I converted to ome.zarr using Fiji. (File -> Save As -> HDF5/OME-NGFF/ZARR)
It is of the format: 3D + T  -> (TZYX)

When opening it using ez-zarr, I get the following error:

```
Cell In[60], line 2
      1 img_directory = 'D://rbc_llsz.ome.zarr/rbc0/'
----> 2 img = ome_zarr.Image(path = img_directory, name = 's0')
      3 img

File c:\Users\Pradeep\.conda\envs\joss_ez_zarr\lib\site-packages\ez_zarr\ome_zarr.py:198, in Image.__init__(self, path, name, skip_checks)
    196 self.pyramid_levels_labels: dict[str, list[str]] = {x: Image._extract_paths_by_decreasing_resolution(self.multiscales_labels[x]['datasets']) for x in self.label_names}
    197 # ... axes units
--> 198 self.axes_unit_image: str = self._load_axes_unit(self.multiscales_image)
    199 self.axes_unit_labels: dict[str, str] = {x: self._load_axes_unit(self.multiscales_labels[x]) for x in self.label_names}
    201 # load channel metadata
    202 # ... label dimensions, e.g. "czyx"

File c:\Users\Pradeep\.conda\envs\joss_ez_zarr\lib\site-packages\ez_zarr\ome_zarr.py:238, in Image._load_axes_unit(multiscale_dict)
    236     return list(unit_set)[0]
    237 else:
--> 238     raise ValueError(f"unsupported unit in multiscale_info: {multiscale_dict}")

ValueError: unsupported unit in multiscale_info: {'name': 'rbc0', 'type': 'Sample', 'version': '0.4', 'axes': [{'type': 'time', 'name': 't', 'unit': 'sec'}, {'type': 'space', 'name': 'z', 'unit': 'micron'}, {'type': 'space', 'name': 'y', 'unit': 'micron'}, {'type': 'space', 'name': 'x', 'unit': 'micron'}], 'datasets': [{'path': 's0', 'coordinateTransformations': [{'scale': [4.752020359039307, 0.3, 0.14499219272808386, 0.14499219272808386], 'type': 'scale'}, {'translation': [0.0, 0.0, 0.0, 0.0], 'type': 'translation'}]}], 'coordinateTransformations': []}
```

Link to the dataset:

[rbc_llsz.ome.zarr.zip](https://github.com/user-attachments/files/19379817/rbc_llsz.ome.zarr.zip)
