### 2025-03-21 fmicompbio/ez_zarr#38 @pr4deepr
working with 3D data
I could have missed this, but I cannot find any documentation in ez_zarr for working with 3D data.
