### 2025-03-21 fmicompbio/ez_zarr#38 @pr4deepr
perhaps you could use something like stackview for interactive sliders?

https://github.com/haesleinhuepf/stackview
