### 2025-03-21 fmicompbio/ez_zarr#38 @pr4deepr
https://github.com/openjournals/joss-reviews/issues/7882
