### 2025-03-21 #7882 @pr4deepr
This is a very useful software tool, considering the increasing popularity of ome zarr file format.
I really liked the ease of use and especially the convenience functions such as adding scale bars, overlaying segmentations. The compatibility with R is definitely a highlight. Although, I have not been able to test the software with R, as I am not familiar with it.

- Could you add code in the documentation on how to generate Figure 1 & 2 of your paper.
- Add a reference to [ngff-zarr ](https://ngff-zarr.readthedocs.io/en/latest/quick_start.html)which includes some functionality to read and write  ome-zarr.
- Is the aim for ez-zarr to be a general purpose tool for ome-zarr imaging data or is it primarily aimed at high content screening? A lot of the focus is on high content screening. If it is aimed to be more generalizable, it would help to provide some examples with other imaging modalities or data types, such as 3D data or 3D+Time, 3D+Channels+Time?
    - Adding this example in the manuscript will make this point clearer as well.
