### 2025-03-25 #7882 @pr4deepr
- Add reference to [napari-ome-zarr-navigator](https://www.napari-hub.org/plugins/) in introduction
