### 2025-03-27 fmicompbio/ez_zarr#38 @mbstadler
`ez_zarr` does in fact support 3D data - it is automatically projecting it to 2D (yx) data using the arguments provided to the `plotting.plot_image` function (possibly via the `ome_zarr.Image.plot` method), in particular `z_projection_method`. This currently supports `'maximum'`, `'minimum'`, `'sum'` and `'average'`. This will work by default if plotting an image with three spatial dimensions.

For special cases, like plotting a single z-plane, or plotting a zy or zx projection or plane, `plotting.zproject` and/or `plotting.plot_image` can be directly called on a pre-processed numpy array.

Thanks for pointing out `stackview`, which would also be easy to use if desired (`img` is an `ome_zarr.Image` object):
```
v = img.get_array_by_coordinate(as_NumPy=True)
stackview.slice(v)
```

We are thinking about how to highlight the 3D support better in the documentation (unfortunately, our sample data has only a single z-plane).
