### 2025-03-27 430eae6
add references to ngff-zarr and napari-ome-zarr-navigator
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
