+++ paper/paper.bib
@@ -145,6 +145,20 @@
   year         = 2025
 }
 
+@MISC{ngff-zarr,
+  title        = "{ngff-zarr}",
+  author       = "McCormick, Matthew",
+  howpublished = "https://pypi.org/project/ngff-zarr/",
+  year         = 2025
+}
+
+@MISC{napari-ome-zarr-navigator,
+  title        = "{napari-ome-zarr-navigator}",
+  author       = "Steffen, Fabio and Luethi, Joel",
+  howpublished = "https://pypi.org/project/napari-ome-zarr-navigator/",
+  year         = 2025
+}
+
 @ARTICLE{harris2020-np,
  title         = {Array programming with {NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
