### 2025-03-27 #7882 @mbstadler
Thank you @pr4deepr for the helpful comments.

Regarding adding the code to generate the figures 1 and 2, which were meant merely for
illustration of the capabilities of `ez_zarr`, we do not think that this would be helpful
because the underlying data is not publicly available, and thus the code would not be runnable.
We believe that we cover all aspects illustrated in these figures also in the examples contained in
our documentation, such as addition of scale bars, zooming, overlaying with label masks etc.

We have added the references to `ngff-zarr` and `napari-ome-zarr-navigator` to the manuscript.

Regarding the scope of the package, you are correct that our use cases were mostly related to
high-content screening, but we think that much of the functionality of `ez_zarr` will also be
helpful in other contexts. For example, we do support 3D images, as described in this
[issue](https://github.com/fmicompbio/ez_zarr/issues/38).
