+++ paper/paper.bib
@@ -148,14 +148,14 @@
 @MISC{ngff-zarr,
   title        = "{ngff-zarr}",
   author       = "McCormick, Matthew",
-  howpublished = "https://pypi.org/project/ngff-zarr/",
+  howpublished = "\url{https://pypi.org/project/ngff-zarr/}",
   year         = 2025
 }
 
 @MISC{napari-ome-zarr-navigator,
   title        = "{napari-ome-zarr-navigator}",
   author       = "Steffen, Fabio and Luethi, Joel",
-  howpublished = "https://pypi.org/project/napari-ome-zarr-navigator/",
+  howpublished = "\url{https://pypi.org/project/napari-ome-zarr-navigator/}",
   year         = 2025
 }
 
