+++ getting_started_Image.qmd
@@ -1,0 +1,507 @@
+---
+jupyter: python3
+title: "Using `ez_zarr` to explore individual OME-Zarr images"
+author: Silvia Barbiero, Michael Stadler, Charlotte Soneson
+toc: true
+---
+
+# Goal
+
+The aim of `ez_zarr` is to provide easy, high-level access to microscopy data, stored in OME-Zarr format according to the [NGFF](https://ngff.openmicroscopy.org/latest/) specifications.
+
+In addition to the python package, we also provide an R package
+that automatically generates and wraps a python environment with
+`ez_zarr` and all dependencies, available at https://github.com/fmicompbio/ezzarr.
+
+The `ome_zarr.Image` class represents an individual image and provides methods to obtain metadata, extract image or label arrays and visualize these. Below we illustrate the use of `ome_zarr.Image` objects by examples.
+
+# Using `ome_zarr.Image` objects
+
+Note: If you prefer to run these examples interactively, you can also download them as an [ipynb notebook](https://fmicompbio.github.io/ez_zarr/getting_started_Image.ipynb).
+
+Here are some examples of how you can use `ome_zarr.Image` objects defined in `ez_zarr`:
+
+## Download example data
+
+To run the code below, you will first need to download some image data.
+You can do this for example by running the following from a terminal:
+```{.bash}
+# Download small Fractal output from Zenodo
+curl -o temp.zip https://zenodo.org/records/10519143/files/20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr.zip
+
+# Unzip and remove the temporary zip file
+unzip temp.zip
+rm temp.zip
+```
+
+or directly from within a python notebook or session using:
+```{python}
+import pandas as pd
+import requests
+import zipfile
+import os
+
+# for info on data see: https://zenodo.org/records/10257149
+url = 'https://zenodo.org/records/10519143/files/20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr.zip'
+
+# download
+response = requests.get(url)
+with open('temp.zip', 'wb') as temp_zip:
+    temp_zip.write(response.content)
+
+# unzip to current directory
+with zipfile.ZipFile('temp.zip', 'r') as zip_ref:
+    zip_ref.extractall('.')
+
+# clean up
+os.remove('temp.zip')
+```
+
+## Load packages
+
+```{python}
+import matplotlib.pyplot as plt
+from matplotlib.colors import ListedColormap
+import numpy as np
+from skimage import color
+
+from ez_zarr import ome_zarr, plotting, utils
+```
+
+## Open OME-Zarr file set
+
+We open a single OME-Zarr image:
+```{python}
+img_directory = '20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr/B/03/0'
+img = ome_zarr.Image(path = img_directory, name = 'cardio-d14')
+img
+```
+
+`img` represents a maximum intensity projection of one well from a high-content screening plate, processed using [Fractal](https://fractal-analytics-platform.github.io/). In addition to such images with two spatial dimensions, `ez-zarr` supports OME-Zarr images with any number of dimensions, automatically projecting images with more than two spatial dimensions for plotting.
+
+In `img`, there are 5 `pyramid_levels` available (0 to 4), meaning that in addition to the full resolution data (level 0), we have four more levels that provide the data in 2-fold lower resolutions (see `pyramid_zyx_scalefactor`), for example for faster plotting.
+
[…]
