### 2025-03-28 fmicompbio/ez_zarr#38 @pr4deepr
Thanks for adding more information.

It would be worth adding a sample 3D data as well at some point.

Not sure if you used this dataset: https://zenodo.org/records/13305316
But, it has both 3D and 2D data.
