### 2025-03-28 fmicompbio/ez_zarr#38 @pr4deepr
Alternatively, a different modality. There are links to a few datasets here:
https://www.openmicroscopy.org/2020/11/04/zarr-data.html

I was using ome-zarr below to download the data:
https://ome-zarr.readthedocs.io/en/stable/cli.html
