### 2025-03-28 #7882 @pr4deepr
Thanks @mbstadler for addressing comments and issues.
I agree with generating code for figures. I misunderstood that and was under the impression you could generate that from a single command.

> Regarding the scope of the package

As the package has mainly been tested on high-content screening data, I think you should reflect that in the manuscript and in the package description.
i.e., The focus is mostly on HCS data or it has been mainly tested with HCS data.  You could add that it should support data from other contexts, but has not been tested extensively.
