### 2025-03-30 22549a4
Add quotes for extras installation
