+++ README.md
@@ -78,7 +78,7 @@ pip install ez-zarr
 
 To install also the dependencies required for running unit tests or to download the example data and run the examples given in the documentation, install `ez_zarr` with the `dev` and `doc` extras, respectively:
 ```
-pip install ez-zarr[dev,doc]
+pip install "ez-zarr[dev,doc]"
 ```
 
 The current (development) `ez_zarr` can be installed from github.com using:
