### 2025-04-02 9416bb9
Expand tests to cover 5d data
