+++ tests/test_ome_zarr.py
@@ -28,6 +28,11 @@ def img2d():
     return ome_zarr.Image('tests/example_data/plate_ones_mip.zarr/B/03/0', name="test")
 
 @pytest.fixture
+def img4d():
+    """A `ome_zarr.Image` object representing a 4D image"""
+    return ome_zarr.Image('tests/example_data/tubhiswt4D_sub.zarr')
+
+@pytest.fixture
 def imgL():
     """A `ome_zarr.ImageList` object with one image"""
     return ome_zarr.ImageList(['tests/example_data/plate_ones_mip.zarr/B/03/0'])
@@ -245,7 +250,7 @@ def test_load_axes_unit(img2d: ome_zarr.Image):
     assert 'organoids' in out_axes_unit_labels
     assert out_axes_unit_labels['organoids'] == 'micrometer'
 
-def test_load_channel_info(img3d: ome_zarr.Image, img2d: ome_zarr.Image):
+def test_load_channel_info(img3d: ome_zarr.Image, img2d: ome_zarr.Image, img4d: ome_zarr.Image):
     """Test `Image._load_channel_info`."""
 
     # input with missing axes
@@ -255,6 +260,7 @@ def test_load_channel_info(img3d: ome_zarr.Image, img2d: ome_zarr.Image):
     # example input
     assert img2d._load_channel_info(img2d.multiscales_image) == 'czyx'
     assert img3d._load_channel_info(img3d.multiscales_image) == 'czyx'
+    assert img4d._load_channel_info(img4d.multiscales_image) == 'tczyx'
     
     # synthetic input
     assert img2d._load_channel_info(
@@ -432,7 +438,7 @@ def test_image_str(img2d: ome_zarr.Image, img3d: ome_zarr.Image):
     assert str(img3d) == repr(img3d)
 
 # ... constructor ...................................................
-def test_constructor(img2d: ome_zarr.Image, img3d: ome_zarr.Image, tmpdir: str):
+def test_constructor(img2d: ome_zarr.Image, img3d: ome_zarr.Image, img4d: ome_zarr.Image, tmpdir: str):
     """Test `ome_zarr.Image` object constructor."""
 
     # non-conforming input
@@ -493,6 +499,22 @@ def test_constructor(img2d: ome_zarr.Image, img3d: ome_zarr.Image, tmpdir: str):
     assert img3d.nchannels_image == 2
     assert isinstance(img3d.channels, list)
     assert all([isinstance(img3d.channels[i], dict) for i in range(img3d.nchannels_image)])
+
+    # 4D image with labels
+    assert isinstance(img4d, ome_zarr.Image)
+    assert img4d.path == 'tests/example_data/tubhiswt4D_sub.zarr'
+    assert img4d.name == 'tubhiswt4D_sub.zarr'
+    assert isinstance(img4d.zarr_group, zarr.Group)
+    assert all([isinstance(x, zarr.Array) for x in img4d.array_dict.values()])
+    assert img4d.ndim == 5
+    assert img4d.label_names == ['embryo']
+    assert img4d.table_names == []
+    assert img4d.axes_unit_image == 'micrometer'
+    assert img4d.channel_info_image == 'tczyx'
+    assert img4d.channel_info_labels == {'embryo': 'tzyx'}
+    assert img4d.nchannels_image == 1
+    assert isinstance(img4d.channels, list)
+    assert all([isinstance(img4d.channels[i], dict) for i in range(img4d.nchannels_image)])
 
 # ... accesssors ....................................................
 def test_get_path(img2d: ome_zarr.Image):
@@ -715,7 +737,7 @@ def test_tree(img2d: ome_zarr.Image):
[…]
