### 2025-04-02 072f442
Add test for too large time index
