### 2025-04-02 56e86c8
Let numpy raise error if time_index is too large
