+++ docs/getting_started_Image.qmd
@@ -39,7 +39,10 @@ or directly from within a python notebook or session using:
 import pandas as pd
 import requests
 import zipfile
+import tempfile
+import shutil
 import os
+import sys
 
 # for info on data see: https://zenodo.org/records/10257149
 url = 'https://zenodo.org/records/10519143/files/20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr.zip'
@@ -421,18 +424,32 @@ img.plot(label_value=877,
 
 ## Calculations on the image data
 
-To save memory, images are stored on disk and only loaded when needed (for implementation details see the [`dask` Array documentation](https://docs.dask.org/en/stable/array.html)).
-
-This can be demonstrated by `type(arr)`:
+Let's obtain again the intensity image data in the form of an array:
 
 ```{python}
-type(arr) # note that the image is an 'on-disk' array
+arr = img.get_array_by_coordinate()
+```
+
+To save memory, images are stored on disk and only loaded when needed (for implementation details see the [`dask` Array documentation](https://docs.dask.org/en/stable/array.html)).
+
+This can be demonstrated by `type(arr)` and `sys.getsizeof(arr)`:
+
+```{python}
+# note that the image is an 'on-disk' array
+type(arr)
+```
+```{python}
+sys.getsizeof(arr)
 ```
 
 To force loading of the data into memory as a `numpy` array, you can call `np.array(arr)`:
 
 ```{python}
-type(np.array(arr)) # triggers loading it into memory
+ndarr = np.array(arr) # triggers loading it into memory
+type(np.array(ndarr))
+```
+```{python}
+sys.getsizeof(ndarr)
 ```
 
 In general, you can use `dask` arrays like `numpy` arrays.
@@ -442,7 +459,8 @@ For example, we can threshold the image:
 # threshold (will also trigger loading into memory)
 arr_mask = arr[0,0]>300
 type(arr_mask)
-
+```
+```{python}
 # plot 
 plotting.plot_image(im=arr[:,0], msk=arr_mask, msk_alpha=0.5,
                     fig_width_inch=6, fig_height_inch=6, fig_dpi=100)
@@ -456,6 +474,63 @@ the recommended way to do more complex tasks that are not directly supported by
 the arguments in `Image.plot`. For example, you could subset a 3D image to select
 just one or several z-planes of interest, instead of using the automated projection
 (`z_projection_method` argument of `plotting.plot_image`).
+
+# Working with images with time dimension
+
+Images that include a time dimension can also be handled using `ez_zarr`. For
+plotting such images, you have to specify which time index to plot.
+
+For illustration, we first download a small example image that is part of
+the `ez_zarr` GitHub repository. It contains free image data showing a
+C. elegans embryo from Josh Bembenek available at
+[openmicroscopy.org](https://docs.openmicroscopy.org/ome-model/5.6.3/ome-tiff/data.html):
+
+```{python}
+# download an image with a time axis
+url = "https://github.com/fmicompbio/ez_zarr/archive/refs/heads/main.zip"
+
+with tempfile.TemporaryDirectory() as temp_dir:
+    response = requests.get(url)
+    response.raise_for_status()  # raise an error if the request fails
+
+    # download
+    temp_zip_file = os.path.join(temp_dir, 'temp.zip')
+    with open(temp_zip_file, 'wb') as temp_zip:
+        temp_zip.write(response.content)
+
+    # unzip to current directory
+    with zipfile.ZipFile(temp_zip_file, 'r') as zip_ref:
+        zip_ref.extractall(temp_dir)
+
+    # locate the extracted folder
+    extracted_folder = os.path.join(temp_dir, "ez_zarr-main", "tests",
[…]
