+++ docs/getting_started_Image.qmd
@@ -508,7 +508,7 @@ with tempfile.TemporaryDirectory() as temp_dir:
 
     # move the image to current working directory
     if os.path.exists(extracted_folder):
-        shutil.copytree(extracted_folder, ".")
+        shutil.copytree(extracted_folder, ".", dirs_exist_ok=True)
 ```
 
 Now we open the image:
