### 2025-04-02 fmicompbio/ez_zarr#38 @mbstadler
We think that would be a useful addition and have added a 5D example data to the package, both for unit tests and documentation. We make use of automated z-projection in the examples added mostly in fb249755b5677bb3d2579ab941b88e766f85ce9d
