### 2025-04-02 fmicompbio/ez_zarr#38 @pr4deepr
Brilliant. Thanks for taking on this feedback.
