### 2025-04-03 fmicompbio/ez_zarr#41 @pr4deepr
spatial transcriptomics data compatibility
I could not get ome-zarr to recognize spatial transcriptomics datasets. Perhaps, I'm using it wrong.

Example datasets:
- [MERFISH technology](https://s3.embl.de/spatialdata/spatialdata-sandbox/merfish.zip)
- [MIBI-TOF](https://s3.embl.de/spatialdata/spatialdata-sandbox/mibitof.zip)

Working on either one of them should be enough to demonstrate compatibility

Related to: https://github.com/openjournals/joss-reviews/issues/7882
