### 2025-04-03 #7882 @pr4deepr
Thanks @mbstadler and @csoneson for taking on feedback.

It makes sense that you'd want to support almost any dataset as long as its an ome-zarr. Adding the example 5D dataset helps a lot. Could you add which modality it was, multiphoton or widefield?

Feel free to include the [dataset I linked](https://github.com/fmicompbio/ez_zarr/issues/37) here as an example. It was taken from this [zenodo repository:](https://doi.org/10.5281/zenodo.7117783). I converted the data: `RBC_tiny.czi` to ome-zarr.

I've added an issue about reading spatial transcriptomics datasets just now.
The 5D+Time and spatial transcriptomics data should cover a broad range of use-cases.
