### 2025-04-03 fmicompbio/ez_zarr#41 @csoneson
Thanks @pr4deepr - actually, if you point `ez-zarr` to the image, you can visualize also these spatial transcriptomics datasets:

```
# MERFISH
img=ome_zarr.Image("data.zarr/images/rasterized")
img.plot()

# MIBI-TOF
img=ome_zarr.Image("data.zarr/images/point8_image")
img.plot()
```

As for the annotations, the issue is similar to #39 - the format is not yet standardized (even these two data sets represent the annotation data differently). Spatial transcriptomics data is definitely one of the application areas that we're keeping a close look at, aiming to support the defined standard as soon as things converge.
