+++ src/ez_zarr/ome_zarr.py
@@ -61,7 +61,8 @@ def create_name_plate_A01(ri: int, ci: int) -> str:
     """
     return f"{chr(ord('A') + ri - 1)}{ci:02}"
 
-def import_plate(path: str, image_name: str = '0') -> "ImageList":
+def import_plate(path: str, image_name: str = '0',
+                 verbose: bool = False) -> "ImageList":
     """
     Create an ImageList object from a OME-Zarr image set corresponding
     to a plate.
@@ -76,6 +77,8 @@ def import_plate(path: str, image_name: str = '0') -> "ImageList":
 
     Parameters:
         path (str): Path to a folder containing the well images.
+        image_name (str): The name of the image to read for each well. 
+        verbose (bool): Whether to print out information messages.
     
     Returns:
         ImageList object
@@ -130,6 +133,8 @@ def import_plate(path: str, image_name: str = '0') -> "ImageList":
     mx_row = max(layout['row_index'])
     mx_column = max(layout['column_index'])
     nrow, ncol = known_plate_dims[min([i for i in range(len(known_plate_dims)) if known_plate_dims[i][0] >= mx_row and known_plate_dims[i][1] >= mx_column])]
+    if verbose:
+        print(f"Identified plate format: {nrow} x {ncol}")
 
     # create ImageList object
     imgL = ImageList(paths = img_paths, names = well_names,
@@ -146,7 +151,8 @@ class Image:
     # constructor and helper functions ----------------------------------------
     def __init__(self, path: str,
                  name: Optional[str]=None,
-                 skip_checks: Optional[bool]=False) -> None:
+                 skip_checks: Optional[bool]=False,
+                 verbose: bool=False) -> None:
         """
         Initializes an OME-Zarr image from its path, containing a
         single zarr group, possibly with multiple resolution levels,
@@ -155,7 +161,8 @@ class Image:
         Parameters:
             path (str): Path containing the OME-Zarr image.
             name (str, optional): Optional name for the image.
-            skip_checks (bool, optional): If `True`, no checks are performed
+            skip_checks (bool, optional): If `True`, no checks are performed.
+            verbose (bool): Whether to print information messages.
         
         Examples:
             Get an object corresponding to an image.
@@ -195,8 +202,8 @@ class Image:
         self.pyramid_levels_image: list[str] = Image._extract_paths_by_decreasing_resolution(self.multiscales_image['datasets'])
         self.pyramid_levels_labels: dict[str, list[str]] = {x: Image._extract_paths_by_decreasing_resolution(self.multiscales_labels[x]['datasets']) for x in self.label_names}
         # ... axes units
-        self.axes_unit_image: str = self._load_axes_unit(self.multiscales_image)
-        self.axes_unit_labels: dict[str, str] = {x: self._load_axes_unit(self.multiscales_labels[x]) for x in self.label_names}
+        self.axes_unit_image: str = self._load_axes_unit(self.multiscales_image, verbose)
+        self.axes_unit_labels: dict[str, str] = {x: self._load_axes_unit(self.multiscales_labels[x], verbose) for x in self.label_names}
 
         # load channel metadata
         # ... label dimensions, e.g. "czyx"
@@ -227,18 +234,21 @@ class Image:
         return info
 
     @staticmethod
-    def _load_axes_unit(multiscale_dict: dict[str, Any]) -> str:
+    def _load_axes_unit(multiscale_dict: dict[str, Any], verbose: bool = False) -> str:
         supported_units = {'micrometer':'micrometer',
                            'micron':'micrometer',
                            'pixel':'pixel',
                            'unit':'pixel'}
[…]
