### 2025-04-03 e2a31ae
Add a few more tests
