### 2025-04-03 1f77035
One more test with verbose=True
