### 2025-04-03 fmicompbio/ez_zarr#41 @pr4deepr
Thanks @csoneson, agree that it is not fully standardized yet, and technologies and formats are evolving rapidly. BTW, I should have clarified that MIBI-TOF is actually spatial proteomics, but your point is valid here.
However, modality specific software, such as [spatialdata ](https://spatialdata.scverse.org/en/latest/tutorials/notebooks/notebooks/examples/technology_merfish.html)is more suited for this I think
Out of curiorsity, in this spatialdata notebook I linked above, is it possible to load points or shapes mentioned here for the MERFISH dataset. They are in geodataframe/shapely formats.
