### 2025-04-04 fmicompbio/ez_zarr#41 @csoneson
> BTW, I should have clarified that MIBI-TOF is actually spatial proteomics, but your point is valid here.

Ah yes, true - but I think we see the same variability also within spatial transcriptomics data sets.

At the moment `ez-zarr` doesn't read `parquet` files. Also here we're watching a bit for the convergence to a tables standard (but agree that conceptually, overlaying quantifications and segmentations also for spatial transcriptomics data would be a use case we'd like to support in the future).
