### 2025-04-04 fmicompbio/ez_zarr#41 @pr4deepr
Thanks for your comment.
I'll close this issue for now..
