### 2025-04-04 77363e8
use ome_zarr instead of hcs_wrappers in __main__.py
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
