+++ src/ez_zarr/__main__.py
@@ -3,10 +3,11 @@
 import sys
 import os
 
-from ez_zarr import hcs_wrappers
+import zarr
+from ez_zarr import ome_zarr
 
 def usage():
-    print('ez_zarr version ' + hcs_wrappers.__version__ + '\n' +
+    print('ez_zarr version ' + ome_zarr.__version__ + '\n' +
           'usage:\n  python -m ez_zarr path\n' +
           'where `path` is a path to an OME-Zarr output from Fractal\n' +
           'or a folder containing such outputs.\n')
@@ -20,10 +21,23 @@ def main():
 
         if os.path.isdir(path):
             path = os.path.normpath(path)
-            if path[-5:] == '.zarr':
-                obj = hcs_wrappers.FractalZarr(path)
+
+            try:
+                zarr_group = zarr.open_group(store=path, mode='r')
+            except Exception as e:
+                print(f"failed to find a zarr group in {path}")
+                sys.exit(1)
+
+            if len([x for x in zarr_group.group_keys() if x not in ['labels', 'tables']]) > 0:
+                try:
+                    obj = ome_zarr.import_plate(path)
+                except Exception as e:
+                    print(f"failed to import plate from {path} - are you pointing at the wrong folder?")
+                    sys.exit(1)
+
             else:
-                obj = hcs_wrappers.FractalZarrSet(path)
+                obj = ome_zarr.Image(path)
+
             print(obj)
         else:
             usage()
