+++ getting_started_hcs_wrappers.qmd
@@ -1,0 +1,459 @@
+---
+jupyter: python3
+title: "Using `ez_zarr` to explore Fractal outputs"
+author: Silvia Barbiero, Michael Stadler
+toc: true
+---
+
+# Disclaimer
+
+The `hcs_wrappers` module described here is only provided for legacy reasons and not being further developed. For new projects, we recommend to use the `ome_zarr` module instead (see [getting started with individual images](https://fmicompbio.github.io/ez_zarr/getting_started_Image/)).
+
+# Goal
+
+The aim of `ez_zarr` is to provide easy, high-level access to high content screening microscopy data, stored as OME-Zarr filesets according to the [NGFF](https://ngff.openmicroscopy.org/latest/) specifications with additional metadata fields, for example the ones generated by the [Fractal](https://fractal-analytics-platform.github.io/) platform.
+
+The goal is that users can write simple scripts working with plates, wells and fields of view, without having to understand how these are represented within an OME-Zarr fileset.
+
+# Installation
+
+`ez_zarr` is hosted on GitHub at [github.com/fmicompbio/ez_zarr](https://github.com/fmicompbio/ez_zarr) and the documentation can be found at [fmicompbio.github.io/ez_zarr](https://fmicompbio.github.io/ez_zarr/).
+
+The release version of `ez_zarr` can be installed using:
+```{.bash}
+pip install ez-zarr
+```
+
+The current (development) version of `ez_zarr` can be installed from github.com using:
+```{.bash}
+pip install git+ssh://git@github.com/fmicompbio/ez_zarr.git
+```
+
+# Quickstart
+
+Note: If you prefer to run these examples interactively, you can also download them as an [ipynb notebook](https://fmicompbio.github.io/ez_zarr/getting_started_hcs_wrappers.ipynb).
+
+Here are some examples of how you can use `ez_zarr`:
+
+## Download example data
+
+To run the code below, you will first need to download some Fractal output data.
+You can do this for example by running the following from a terminal:
+```{.bash}
+# Download small Fractal output from Zenodo
+curl -o temp.zip https://zenodo.org/records/10519143/files/20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr.zip
+
+# Unzip and remove the temporary zip file
+unzip temp.zip
+rm temp.zip
+```
+
+or directly from within a python notebook or session using:
+```{python}
+import requests
+import zipfile
+import os
+
+# for info on data see: https://zenodo.org/records/10257149
+url = 'https://zenodo.org/records/10519143/files/20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr.zip'
+
+# download
+response = requests.get(url)
+with open('temp.zip', 'wb') as temp_zip:
+    temp_zip.write(response.content)
+
+# unzip to current directory
+with zipfile.ZipFile('temp.zip', 'r') as zip_ref:
+    zip_ref.extractall('.')
+
+# clean up
+os.remove('temp.zip')
+```
+
+## Load packages
+
+```{python}
+import matplotlib.pyplot as plt
+from matplotlib.colors import ListedColormap
+import numpy as np
+from skimage import color
+
+from ez_zarr import hcs_wrappers, plotting
+```
+
+## Open OME-Zarr file set
+
+We open a single OME-Zarr directory, typically representing a single plate:
+```{python}
+zarr_directory = '20200812-CardiomyocyteDifferentiation14-Cycle1_mip.zarr'
+plate = hcs_wrappers.FractalZarr(zarr_directory, name = 'cardio-d14')
+plate
+```
+
+`plate` represents a maximum intensity projection of an acquisition with 1 well and 1 channel.
+
[…]
