### 2025-05-02 fmicompbio/ez_zarr#45 @ianhi
Document Supported Versions of OME-Zarr Spec
There are multiple versions of the OME-Zarr spec: https://ngff.openmicroscopy.org/specifications/index.html

In testing I found that there is an unhandled exception when loading data saved in the `0.1` spec:

```python
from ez_zarr import ome_zarr
im = ome_zarr.Image(path="https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.1/4007801.zarr")
```

Gives:

```python
IndexError                                Traceback (most recent call last)
Cell In[58], line 2
      1 from ez_zarr import ome_zarr
----> 2 im = ome_zarr.Image(path="https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.1/4007801.zarr")

File [~/miniforge3/envs/ez-zarr/lib/python3.12/site-packages/ez_zarr/ome_zarr.py:175](http://localhost:8890/lab/tree/~/miniforge3/envs/ez-zarr/lib/python3.12/site-packages/ez_zarr/ome_zarr.py#line=174), in Image.__init__(self, path, name, skip_checks)
    173 self.zarr_group: zarr.Group = zarr.open_group(store=self.path, mode='r')
    174 self.array_dict: dict[str, zarr.Array] = {x[0]: x[1] for x in self.zarr_group.arrays()}
--> 175 self.ndim = self.array_dict[list(self.array_dict.keys())[0]].ndim
    176 self.label_names = []
    177 if 'labels' in list(self.zarr_group.group_keys()):

IndexError: list index out of range
```

despite being a valid ome-zarr file: https://ome.github.io/ome-ngff-validator/?source=https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.1/1884807.zarr

## Potential Solution

It's not necessary to support every version of OME-Zarr, but as a user I would expect that which versions are supported to be stated clearly and early in the documentation.

Ideally also failing gracefully if encountering an ome-zarr file with a spec version not supported.
