### 2025-05-02 fmicompbio/ez_zarr#45 @ianhi
This may also apply to the tables from fractal
