### 2025-05-02 #7882 @ianhi
Hi all,

I'm very sorry for the delays on my end. I've filled out my review above. I've currently left one box unchecked (should be an easy fix) for the reason described below:

# Functionality

I've checked every box except for the functionality as the software does not currently make note of which versions of the OME-Zarr spec it can read. Without this it is easy for a user to read the claim as reading any and all OME-Zarr files which as I note in the relevant issue is not true: https://github.com/fmicompbio/ez_zarr/issues/45

I particularly appreciated that there are R bindings. This is an excellent step for increasing usage of a common file format.

# Documentation
The main example is useful and seems comprehensive. Going forward if they want to support more diverse use cases it might be nice to see more examples of datasets with different shapes. e.g. something from https://idr.github.io/ome-ngff-samples/

However, overall the documentation is thorough and readable.

# Paper

This is a nice and too the point paper. The only thing that might be considered missing is any mention of `Fractal` and the support `ez_zarr` has for fractal tables. Something that I don't think any other software can claim. This isn't necessary to include, but given that the software does support it I'd encourage adding a note about that.
