### 2025-05-04 d9e68f5
Check OME-Zarr version, document supported version
