### 2025-05-06 #7882 @mbstadler
Thank you @ianhi for the constructive feedback!

We have implemented some edits (https://github.com/fmicompbio/ez_zarr/pull/46) in response to your issue concerning supported OME-Zarr versions.

Regarding Fractal tables, we thank you for appreciating our current support in `ez-zarr`. However, we would prefer to not discuss tables at this point in detail, since they are not yet part of the OME-Zarr standard, and the current implementation of tables in Fractal is still dynamic and likely to change. We thus consider our table support preliminary and also likely subject to change.

Please let us know if there is anything left for us to do, or any open issues from your side.
