### 2025-05-06 fmicompbio/ez_zarr#45 @ianhi
Nice
