### 2025-05-06 #7882 @ianhi
I see what you mean regarding fractal

> Please let us know if there is anything left for us to do, or any open issues from your side.

Nothing from me! I've checked every box now. Thanks for making software in this space :) bioimage loading tools are super useful
