### 2025-05-09 a9efa2f
bump version (JOSS accepted version)
Co-authored-by: Charlotte Soneson <charlottesoneson@gmail.com>
Co-authored-by: Michael Stadler <stadler.michael@gmail.com>
