+++ src/ez_zarr/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '0.3.7'
+__version__ = '0.3.8'
 __author__ = 'Silvia Barbiero, Michael Stadler, Charlotte Soneson'
