+++ src/ez_zarr/ome_zarr.py
@@ -9,7 +9,7 @@ Classes:
 
 __all__ = ['Image', 'ImageList', 'create_name_row_col', 
            'create_name_plate_A01', 'import_plate']
-__version__ = '0.3.7'
+__version__ = '0.3.8'
 __author__ = 'Silvia Barbiero, Michael Stadler, Charlotte Soneson'
 
 
