+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "ez_zarr"
-version = "0.3.7"
+version = "0.3.8"
 description = "Give easy, high-level access to ome-zarr filesets."
 readme = { file = "README.md", content-type = "text/markdown" }
 authors = [
