### 2025-05-09 #7882 @mbstadler
Thank you to all the reviewers and @adamltyson!

We have:

- [x] double checked authors, affiliations and ORCIDs
- [x] made a release of the software ([v0.3.8](https://github.com/fmicompbio/ez_zarr/releases/tag/v0.3.8)) that includes all the changes that were implemented during the review
- [x] archived this version on Zenodo: [10.5281/zenodo.15370320](https://doi.org/10.5281/zenodo.15370320)
- [x] made sure that title, author list (including ORCIDs) in the Zenodo archive match those in the JOSS paper
- [x] made sure that the license and copyright listed in the Zenodo archive match those in the repository

Please let us know if there is anything missing from our side.
