### 2025-05-17 #7882 @Kevin-Mattheus-Moerman
- [ ] The paper features European/British English (e.g. colour, visualised etc) but the spelling of `specialized` is American English. Please consider writing that as `specialised` for consistency.
