+++ paper/paper.md
@@ -54,7 +54,7 @@ Finally, while `ez-zarr` is implemented in python and installable using `pip` an
 
 OME-Zarr is a relatively new file format and not yet supported by many image analysis tools. There are tools that allow accessing OME-Zarr files, amongst them interactive viewers, such as `napari` [@Sofroniew2024-cz;@napari-ome-zarr-navigator] and `MoBIE` [@Pape2023-hr], and low level libraries to read and write OME-Zarrs, including `ome-zarr` [@ome-zarr], `bioio-ome-zarr` [@bioio], `ngio` [@ngio], `ngff-zarr` [@ngff-zarr] and `zarr` [@zarr].
 
-While these tools allow interactive exploration of images in OME-Zarr format, or getting low-level access to the contained data, we were not aware of tools that provide higher-level access and scriptable visualisation capabilities. Some more specialized projects also cover such tasks, but focus on specific types of imaging data, for example spatial omics data in the `SpatialData` framework [@Marconato2024-ie]. Furthermore, most available tools are implemented in python and do not have an equivalent interface in other languages like R. `ez-zarr` was designed to fill these gaps, building on the low-level file access functionality from the `zarr` library, and provide a minimal, abstract interface to OME-Zarr images or collections of such images from both python and R.
+While these tools allow interactive exploration of images in OME-Zarr format, or getting low-level access to the contained data, we were not aware of tools that provide higher-level access and scriptable visualisation capabilities. Some more specialised projects also cover such tasks, but focus on specific types of imaging data, for example spatial omics data in the `SpatialData` framework [@Marconato2024-ie]. Furthermore, most available tools are implemented in python and do not have an equivalent interface in other languages like R. `ez-zarr` was designed to fill these gaps, building on the low-level file access functionality from the `zarr` library, and provide a minimal, abstract interface to OME-Zarr images or collections of such images from both python and R.
 
 `ez-zarr` has been used productively by several research groups and in published research [@Schwayer2025]. 
 
