### 2025-05-17 #7882 @csoneson
Thanks @Kevin-Mattheus-Moerman - the spelling of specialised has been [fixed](https://github.com/fmicompbio/ez_zarr/commit/8bd7855d5f4f3d2297066e7698b77955f46fd263).
